Migrate AdvancedProtectionStatusManager to the IdentityManager
This CL moves AdvancedProtectionStatusManager away from using SiginManager
and related APIs (e.g. SiginManagerBase::Observer), replacing those bits
with calls to the IdentityManager's API instead.

Likewise it moves from fetching tokens using the PO2TS APIs to building
a PrimaryAccessTokenFetcher and relying on it to retrieve the tokens
associated with the primary account.

Bug: 887264