Simplify token retrieval logic in TokenHandleFetcher
Instead of relying on an implementation of IdentityManager::Observer and
a private flag to know when a RefreshToken is available before requesting
using PrimaryAccountAccessTokenFetcher, use PAATF's mode to wait until the
primary account has a refresh token available (i.e. kWaitUntilAvailable)
whenever needed when TokenHandleFetcher::BackfillToken() is invoked, or
simply using kImmediate if such refresh token is already available.

Bug: 905689