Implement legacy methods in PrimaryAccountMutator related to signin process
Implement the following methods at the same time, so that we can also
provide decent unit tests that would require of all of them to make
sure functionality works as expected:

  * LegacyStartSigninWithRefreshTokenForPrimaryAccount()
  * LegacyCompletePendingPrimaryAccountSignin()
  * LegacyIsPrimaryAccountAuthInProgress()
  * LegacyPrimaryAccountForAuthInProgress()

Also, PrimaryAccountMutatorTests unit tests for each of those methods as
part of this CL, based on those from SigninManagerTest.

TODO: Implement tests for LegacyStartSigninWithRefreshTokenForPrimaryAccount
and LegacyCompletePendingPrimaryAccountSignin.

Bug: 906063