Ensure SigninManager::Initialize() always receives a non-null PrefService*
Check that the passed |PrefService* local_state| parameter is not null
and fix all related tests to make sure that instances of SigninManager
can always receive a valid local_state from unit tests.

To achieve this, we need to ensure that such unit tests do always
register a valid PrefService* as the local state of the browser
process before SigninManager::Initialize() gets called, and that
such local state in the browser process gets restored when needed.

Bug: 908121